requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
django-cors-headers==4.3.1
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
requests==2.31.0
